app/requirements.txt
# Database
psycopg2-binary>=2.9.10
# HTTP & APIs
requests==2.31.0
anthropic==0.18.1
# Telegram
python-telegram-bot>=21.9
# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1
# Utilities
pytz==2024.1
# Testing
pytest==7.4.3
pytest-cov==4.1.0
